fix option import: No results from Akeneo
Currently, if the first option which has a valid type has no options, the whole import is terminated. This happens because the foreach is terminated after the first attribute with valid type.
With the fix, the foreach is continued until there is an attribute with options or the foreach has finished.